Module Atdgen_runtime.Json_adapter

module type S = sig ... end

Module signature required of any json adapter. For example, an ATD annotation <json adapter.ocaml="Atdgen_runtime.Json_adapter.Type_field" refers to the OCaml module Atdgen_runtime.Json_adapter.Type_field.

module Type_field : sig ... end

Support for json objects that contain a field that indicates the type of that object. The following

module One_field : S

Support for objects with a single field whose name indicates the type of the value. For instance,

module Type_and_value_fields : sig ... end

Support for the retired tag_field feature. This converts the following