Parameter Make.1-Param

val type_field_name : string

The name of the json field containing the type of the value as a json string.

val value_field_name : string

The name of the json field containing the value.

val known_tags : (string list * string) option

Optionally indicate a set of known tags and a catch-all tag. This allows unknown tags/values coming from the original json to be wrapped under a catch-all tag.